Move ApplicationController logic to concerns
